Split conversion (ease of future change)
